Major refactoring. Use computerscreen instead of LED matrix.
Created abstract base classes for animation and display.
Refactored gameframe animation to adopt the new style. Not finished yet.
Added the possibility to output everything on the local computer
display. That makes development hopefully easier. Pygame is used here.
Pygame must be used on mainthread. That's why I finally started to work
on the clue and "brain": ribbapi.py. There is a queue so that animation
threads can put the current frame on the queue for the mainthread to
consume (display).
Many things have not been adopted yet. And of course "the brain" needs a
lot more work.